+++ Mud.Server.Interfaces/Ability/IRegenerationPassive.cs
@@ -5,6 +5,6 @@ namespace Mud.Server.Interfaces.Ability;
 
 public interface IRegenerationPassive : IPassive
 {
-    int HitGainModifier(ICharacter user, int baseHitGain);
-    int ResourceGainModifier(ICharacter user, ResourceKinds resourceKind, int baseResourceGain);
+    decimal HitGainModifier(ICharacter user, decimal baseHitGain);
+    decimal ResourceGainModifier(ICharacter user, ResourceKinds resourceKind, decimal baseResourceGain);
 }
